move network selector and add wallet feature into separate view
